Add files via upload
BGtopembed are the csv files containing the correct labelling along with the wrong labelling of the Bhagvad Gita question answer dataset. These files contain the labels 0 for correct pair and 1 for wrong pairs. 

The top embeding followed by numbers represent the rank of wrong question that was associated with the given answer using the all-mini-V6-LM SBERT model.